Menu: allow nav without mouse
Allows keyboard or controller navigation of the menu system.

1. EventSystem - Set "Force Module Active = True" in Standalone Input
Module ... this causes the navigation system to be active immediately.
2. Choose a specific button to be the first one selected (for us it's
the "PlayButton" on the MainMenu.
3. When clicking a button, not only do we change the active GameObject,
but we also will select a specific menu button on that page. For
example, when navigating from `MainMenu => CharacterSelect`, we set
the `NextButton` to be selected on the CharacterSelect page.