support character select via 'lobby' UX
This is a work in progress, but hopefully with just a little more
cleanup will be mergeable.

I would like to refactor very classes to simplify event passing and
reduce the number of interdependencies (spaghetti code alert!).

The core ideas are:
- human players should only join/leave within the CharacterSelect UI
- during character select, humans should only be able to select their on
character
- whatever state/configuration you set on CharacterSelect should be
passed into the gameplay scene
- any human player should be able to navigate the UI on other pages.
Just the character select 'lobby' is special.

Challenges include:
- Passing PlayerInput (humans with controllers) from the Menu scene to
the gameplay Levels.
	- I did this by creating a HumanPlayer concept and passing it
via DontDestroyOnLoad. The HumanPlayer is then responsible for
controlling and spawning whatever is needed within various UIs.. For
example setting the right controller inputs, spawning an in-game avatar,
etc.
- Moving between UI Input and PlayerInput.
	- UI input needs to reason about the currently selected UI
element
	- player input sometimes takes the same controller key as input
(e.g. start), but this can double-trigger to click a UI input (clicking
a button on the subsequent page with UI elements) in some cases.
- Not writing crappy code. Ends up I haven't written a lot of truly
object-oriented code, and I'm feeling the pains!
- `GameObject.Find()` seems risky, since whatever you're trying to find
might not exist and you need to check for nulls.. But if you have
Prefabs, this is the best you can do. Prefabs cannot have references to
actual in game instances (which makes sense-- they aren't in the scene
until you instantiate them!).