actions: split up build + test
- only run `test` pre-merge
- run builds on master

Why? It's very slow to run all the builds on the PR branch (15m for web
build). We have ~1000m/month for Github actions free, and I think this
will eat them up.

Potentially we can do one build on the branch if that gives us
additional confidence.

Having builds on master is most important so we have releases we can all
be running / testing / comparing. E.g. it will be easy to file a bug
against a given commit.