fix wraparound bug
Before: The wraparound logic was pulling you slightly "out of bounds"
instead of pushing you back "in bounds". This meant that when you
wrapped around, it was possible to fall outside the level, because you
wouldn't retrigger the boundary.

Now: you shouldn't ever be able to get outside the boundary. You should
always hit the trigger and wraparound successfully.